+++ localbuild/meta.yaml
@@ -19,6 +19,7 @@ build:
     - mswms_demodata = mslib.mswms.demodata:main
     - mscolab = mslib.mscolab.mscolab:main
     - mssautoplot = mslib.utils.mssautoplot:main
+    - msidp = mslib.msidp.idp:main
 
 requirements:
   build:
@@ -104,6 +105,7 @@ test:
     - mswms_demodata -h
     - msui -h
     - mscolab -h
+    - msidp -h
 
 about:
   summary: 'A web service based tool to plan atmospheric research flights.'
