+++ mslib/mscolab/mscolab.py
@@ -34,10 +34,11 @@ import sys
 import secrets
 import subprocess
 import time
+import git
 
 from mslib import __version__
 from mslib.mscolab.conf import mscolab_settings
-from mslib.mscolab.seed import seed_data, add_user, add_all_users_default_operation,\
+from mslib.mscolab.seed import seed_data, add_user, add_all_users_default_operation, \
     add_all_users_to_all_operations, delete_user
 from mslib.mscolab.utils import create_files
 from mslib.utils import setup_logging
@@ -94,32 +95,41 @@ def handle_db_seed():
     seed_data()
     print("Database seeded successfully!")
 
+
 def handle_mscolab_certificate_init():
     print('generating CRTs for the mscolab server......')
 
     try:
-        cmd = f"openssl req -newkey rsa:4096 -keyout {mscolab_settings.MSCOLAB_SSO_DIR}/key_mscolab.key -nodes -x509 -days 365 -batch -subj '/CN=localhost' -out {mscolab_settings.MSCOLAB_SSO_DIR}/crt_mscolab.crt"
-        subprocess.run(cmd, shell=True, check=True)
+        cmd = ["openssl", "req", "-newkey", "rsa:4096", "-keyout",
+               f"{mscolab_settings.MSCOLAB_SSO_DIR}/key_mscolab.key",
+               "-nodes", "-x509", "-days", "365", "-batch", "-subj",
+               "/CN=localhost", "-out", f"{mscolab_settings.MSCOLAB_SSO_DIR}/crt_mscolab.crt"]
+        subprocess.run(cmd, check=True)
         print("generated CRTs for the mscolab server.")
         return True
     except subprocess.CalledProcessError as error:
         print(f"Error while generating CRTs for the mscolab server: {error}")
         return False
 
+
 def handle_local_idp_certificate_init():
     print('generating CRTs for the local identity provider......')
 
     try:
-        cmd = f"openssl req -newkey rsa:4096 -keyout {mscolab_settings.MSCOLAB_SSO_DIR}/key_local_idp.key -nodes -x509 -days 365 -batch -subj '/CN=localhost' -out {mscolab_settings.MSCOLAB_SSO_DIR}/crt_local_idp.crt"
-        subprocess.run(cmd, shell=True, check=True)
+        cmd = ["openssl", "req", "-newkey", "rsa:4096", "-keyout",
+               f"{mscolab_settings.MSCOLAB_SSO_DIR}/key_local_idp.key",
+               "-nodes", "-x509", "-days", "365", "-batch", "-subj",
+               "/CN=localhost", "-out", f"{mscolab_settings.MSCOLAB_SSO_DIR}/crt_local_idp.crt"]
+        subprocess.run(cmd, check=True)
         print("generated CRTs for the local identity provider")
         return True
     except subprocess.CalledProcessError as error:
         print(f"Error while generated CRTs for the local identity provider: {error}")
         return False
 
+
 def handle_mscolab_backend_yaml_init():
-    saml_2_backend_yaml_content ="""name: Saml2
+    saml_2_backend_yaml_content = """name: Saml2
 config:
   entityid_endpoint: true
   mirror_force_authn: no
@@ -176,7 +186,8 @@ config:
         endpoints:
           assertion_consumer_service:
             - [http://localhost:8083/localhost_test_idp/acs/post, 'urn:oasis:names:tc:SAML:2.0:bindings:HTTP-POST']
-            - [http://localhost:8083/localhost_test_idp/acs/redirect, 'urn:oasis:names:tc:SAML:2.0:bindings:HTTP-Redirect']
+            - [http://localhost:8083/localhost_test_idp/acs/redirect,
+            'urn:oasis:names:tc:SAML:2.0:bindings:HTTP-Redirect']
           discovery_response:
           - [<base_url>/<name>/disco, 'urn:oasis:names:tc:SAML:profiles:SSO:idp-discovery-protocol']
         name_id_format: 'urn:oasis:names:tc:SAML:2.0:nameid-format:transient'
@@ -238,33 +249,36 @@ config:
   #       name_id_format_allow_create: true
 """
     try:
-        file_path=f"{mscolab_settings.MSCOLAB_SSO_DIR}/mss_saml2_backend.yaml"
+        file_path = f"{mscolab_settings.MSCOLAB_SSO_DIR}/mss_saml2_backend.yaml"
         with open(file_path, "w", encoding="utf-8") as file:
             file.write(saml_2_backend_yaml_content)
         return True
-    except (FileNotFoundError,PermissionError) as error:
+    except (FileNotFoundError, PermissionError) as error:
         print(f"Error while generated backend .yaml for the local mscolabserver: {error}")
         return False
 
-def handle_mscolab_metadata_init():
+
+def handle_mscolab_metadata_init(repo_exists):
     '''
         This will generate necessary metada data file for sso in mscolab through localhost idp
 
         Before running this function:
         - Ensure that USE_SAML2 is set to True.
-        - Generate the necessary keys and certificates and configure them in the .yaml 
+        - Generate the necessary keys and certificates and configure them in the .yaml
         file for the local IDP.
     '''
     print('generating metadata file for the mscolab server')
 
     try:
-        process = subprocess.Popen(["python", "mslib/mscolab/mscolab.py", "start"])
+        command = ["python", "mslib/mscolab/mscolab.py", "start"] if repo_exists else ["mscolab", "start"]
+        process = subprocess.Popen(command)
 
         # Add a small delay to allow the server to start up
         time.sleep(10)
 
-        cmd_curl = f"curl http://localhost:8083/metadata/ -o {mscolab_settings.MSCOLAB_SSO_DIR}/metadata_sp.xml"
-        subprocess.run(cmd_curl, shell=True, check=True)
+        cmd_curl = ["curl", "http://localhost:8083/metadata/",
+                    "-o", f"{mscolab_settings.MSCOLAB_SSO_DIR}/metadata_sp.xml"]
+        subprocess.run(cmd_curl, check=True)
         process.kill()
         print('mscolab metadata file generated succesfully')
         return True
@@ -274,14 +288,25 @@ def handle_mscolab_metadata_init():
         return False
 
 
-def handle_local_idp_metadata_init():
+def handle_local_idp_metadata_init(repo_exists):
     print('generating metadata for localhost identity provider')
 
     try:
         if os.path.exists(f"{mscolab_settings.MSCOLAB_SSO_DIR}/idp.xml"):
             os.remove(f"{mscolab_settings.MSCOLAB_SSO_DIR}/idp.xml")
-        cmd = f"make_metadata mslib/idp/idp_conf.py > {mscolab_settings.MSCOLAB_SSO_DIR}/idp.xml"
-        subprocess.run(cmd, shell=True, check=True)
+
+        idp_conf_path = "mslib/msidp/idp_conf.py"
+
+        if not repo_exists:
+            import site
+            site_packages_path = site.getsitepackages()[0]
+            idp_conf_path = os.path.join(site_packages_path, "mslib/msidp/idp_conf.py")
+
+        cmd = ["make_metadata", idp_conf_path]
+
+        with open(f"{mscolab_settings.MSCOLAB_SSO_DIR}/idp.xml",
+                  "w", encoding="utf-8") as output_file:
+            subprocess.run(cmd, stdout=output_file, check=True)
         print("idp metadata file generated succesfully")
         return True
     except subprocess.CalledProcessError as error:
@@ -290,6 +315,7 @@ def handle_local_idp_metadata_init():
             os.remove(f"{mscolab_settings.MSCOLAB_SSO_DIR}/idp.xml")
         print(f"Error while generating metadata for localhost identity provider: {error}")
         return False
+
 
 def handle_sso_crts_init():
     """
@@ -314,18 +340,17 @@ def handle_sso_crts_init():
     print('\n\nAll CRTs and mscolab backend saml files generated successfully !')
 
 
-def handle_sso_metadata_init():
+def handle_sso_metadata_init(repo_exists):
     print('\n\ngenerating metadata files.......')
-    if not handle_mscolab_metadata_init():
+    if not handle_mscolab_metadata_init(repo_exists):
         print('Error while handling mscolab metadata.')
         return
 
-    if not handle_local_idp_metadata_init():
+    if not handle_local_idp_metadata_init(repo_exists):
         print('Error while handling idp metadata.')
         return
-    
+
     print("\n\nALl necessary metadata files generated successfully")
-    
 
 
 def main():
@@ -356,9 +381,12 @@ def main():
                                  action="store_true")
     sso_conf_parser = subparsers.add_parser("sso_conf", help="single sign on process configurations")
     sso_conf_parser = sso_conf_parser.add_mutually_exclusive_group(required=True)
-    sso_conf_parser.add_argument("--init_sso_crts",help="Generate all the essential CRTs required for the Single Sign-On process using the local Identity Provider",
+    sso_conf_parser.add_argument("--init_sso_crts",
+                                 help="Generate all the essential CRTs required for the Single Sign-On process "
+                                 "using the local Identity Provider",
                                  action="store_true")
-    sso_conf_parser.add_argument("--init_sso_metadata",help="Generate all the essential metadata files required for the Single Sign-On process using the local Identity Provider",
+    sso_conf_parser.add_argument("--init_sso_metadata", help="Generate all the essential metadata files required "
+                                 "for the Single Sign-On process using the local Identity Provider",
                                  action="store_true")
 
     args = parser.parse_args()
@@ -370,6 +398,13 @@ def main():
         print("Documentation: http://mss.rtfd.io")
         print("Version:", __version__)
         sys.exit()
+
+    try:
+        _ = git.Repo(os.path.dirname(os.path.realpath(__file__)), search_parent_directories=True)
+        repo_exists = True
+
+    except git.exc.InvalidGitRepositoryError:
+        repo_exists = False
 
     updater = Updater()
     if args.update:
@@ -430,7 +465,8 @@ def main():
     elif args.action == "sso_conf":
         if args.init_sso_crts:
             confirmation = confirm_action(
-                "This will reset and initiation all CRTs and SAML yaml file as default. Are you sure to continue? (y/[n]):")
+                "This will reset and initiation all CRTs and SAML yaml file as default. "
+                "Are you sure to continue? (y/[n]):")
             if confirmation is True:
                 handle_sso_crts_init()
         if args.init_sso_metadata:
@@ -438,17 +474,20 @@ def main():
                 "Are you sure you executed --init_sso_crts before running this? (y/[n]):")
             if confirmation is True:
                 confirmation = confirm_action(
-                """
-                This will generate necessary metada data file for sso in mscolab through localhost idp
+                    """
+                    This will generate necessary metada data file for sso in mscolab through localhost idp
 
-                Before running this function:
-                - Ensure that USE_SAML2 is set to True.
-                - Generate the necessary keys and certificates and configure them in the .yaml 
-                file for the local IDP.
-                
-                Are you sure you set all correctly as per the documentation? (y/[n]):""")
+                    Before running this function:
+                    - Ensure that USE_SAML2 is set to True.
+                    - Generate the necessary keys and certificates and configure them in the .yaml
+                    file for the local IDP.
+
+                    Are you sure you set all correctly as per the documentation? (y/[n]):
+                    """
+                )
                 if confirmation is True:
-                    handle_sso_metadata_init()
+                    handle_sso_metadata_init(repo_exists)
+
 
 if __name__ == '__main__':
     main()
