+++ mslib/mscolab/server.py
@@ -30,8 +30,6 @@ import logging
 import time
 import datetime
 import secrets
-import random
-import string
 import warnings
 import sys
 import fs
@@ -52,6 +50,7 @@ from saml2.config import SPConfig
 from saml2.client import Saml2Client
 from saml2.metadata import create_metadata_string
 from saml2 import BINDING_HTTP_REDIRECT, BINDING_HTTP_POST, SAMLError
+from flask.wrappers import Response
 
 from mslib.mscolab.conf import mscolab_settings
 from mslib.mscolab.models import Change, MessageType, User, Operation, db
@@ -60,7 +59,6 @@ from mslib.mscolab.utils import create_files, get_message_dict
 from mslib.utils import conditional_decorator
 from mslib.index import create_app
 from mslib.mscolab.forms import ResetRequestForm, ResetPasswordForm
-from flask.wrappers import Response
 
 APP = create_app(__name__)
 mail = Mail(APP)
@@ -80,8 +78,8 @@ except ImportError as ex:
                          ("add_new_user_here", "add_md5_digest_of_PASSWORD_here")]
         __file__ = None
 
-#setup idp login config
-if mscolab_settings.USE_SAML2 :
+# setup idp login config
+if mscolab_settings.USE_SAML2:
     with open(f"{mscolab_settings.MSCOLAB_SSO_DIR}/mss_saml2_backend.yaml", encoding="utf-8") as fobj:
         yaml_data = yaml.safe_load(fobj)
 
@@ -89,9 +87,12 @@ if mscolab_settings.USE_SAML2 :
     for configured_idp in mscolab_settings.CONFIGURED_IDPS:
         # set CRTs and metadata paths for the localhost_test_idp
         if 'localhost_test_idp' in configured_idp['idp_identity_name']:
-            yaml_data["config"]["localhost_test_idp"]["key_file"] = f'{mscolab_settings.MSCOLAB_SSO_DIR}/key_mscolab.key'
-            yaml_data["config"]["localhost_test_idp"]["cert_file"] = f'{mscolab_settings.MSCOLAB_SSO_DIR}/crt_mscolab.crt'
-            yaml_data["config"]["localhost_test_idp"]["metadata"]["local"][0] = f'{mscolab_settings.MSCOLAB_SSO_DIR}/idp.xml'
+            yaml_data["config"]["localhost_test_idp"]["key_file"] = \
+                f'{mscolab_settings.MSCOLAB_SSO_DIR}/key_mscolab.key'
+            yaml_data["config"]["localhost_test_idp"]["cert_file"] = \
+                f'{mscolab_settings.MSCOLAB_SSO_DIR}/crt_mscolab.crt'
+            yaml_data["config"]["localhost_test_idp"]["metadata"]["local"][0] = \
+                f'{mscolab_settings.MSCOLAB_SSO_DIR}/idp.xml'
 
     if not os.path.exists(yaml_data["config"]["localhost_test_idp"]["metadata"]["local"][0]):
         yaml_data["config"]["localhost_test_idp"]["metadata"]["local"] = []
@@ -231,22 +232,6 @@ def verify_user(func):
                 return func(*args, **kwargs)
     return wrapper
 
-# ToDo refactor, have also a look on secrets? see discussion 
-# in https://github.com/Open-MSS/MSS/pull/1818#discussion_r1270701658
-def rndstr(size=16, alphabet=""):
-    """
-    Returns a string of random ascii characters or digits
-    :type size: int
-    :type alphabet: str
-    :param size: The length of the string
-    :param alphabet: A string with characters.
-    :return: string
-    """
-    rng = random.SystemRandom()
-    if not alphabet:
-        alphabet = string.ascii_letters[0:52] + string.digits
-    return type(alphabet)().join(rng.choice(alphabet) for _ in range(size))
-
 
 def get_idp_entity_id(selected_idp):
     """
@@ -269,7 +254,6 @@ def get_idp_entity_id(selected_idp):
     return entity_id
 
 
-
 @APP.route('/')
 def home():
     return render_template("/index.html")
@@ -279,8 +263,10 @@ def home():
 @conditional_decorator(auth.login_required, mscolab_settings.__dict__.get('enable_basic_http_authentication', False))
 def hello():
     return json.dumps({
-            'message': "Mscolab server",
-            'USE_SAML2': mscolab_settings.USE_SAML2})
+        'message': "Mscolab server",
+        'USE_SAML2': mscolab_settings.USE_SAML2
+    })
+
 
 @APP.route('/token', methods=["POST"])
 @conditional_decorator(auth.login_required, mscolab_settings.__dict__.get('enable_basic_http_authentication', False))
@@ -776,6 +762,7 @@ def reset_request():
         logging.warning("To send emails, the value of `MAIL_ENABLED` in `conf.py` should be set to True.")
         return render_template('errors/403.html'), 403
 
+
 @APP.route("/metadata/", methods=['GET'])
 def metadata():
     """Return the SAML metadata XML for congiguring local host testing IDP"""
@@ -783,6 +770,7 @@ def metadata():
         None, sp_localhost_test_idp.config, 4, None, None, None, None, None
     ).decode("utf-8")
     return Response(metadata_string, mimetype="text/xml")
+
 
 @APP.route('/available_idps/', methods=['GET'])
 def available_idps():
@@ -793,7 +781,7 @@ def available_idps():
     """
     if mscolab_settings.USE_SAML2:
         configured_idps = mscolab_settings.CONFIGURED_IDPS
-        return render_template('idp/available_idps.html', configured_idps=configured_idps), 200   
+        return render_template('idp/available_idps.html', configured_idps=configured_idps), 200
     return render_template('errors/403.html'), 403
 
 
@@ -814,12 +802,10 @@ def idp_login():
         _, response_binding = sp_config.config.getattr("endpoints", "sp")[
             "assertion_consumer_service"
         ][0]
-        relay_state = rndstr()
         entity_id = get_idp_entity_id(selected_idp)
         _, binding, http_args = sp_config.prepare_for_negotiated_authenticate(
             entityid=entity_id,
             response_binding=response_binding,
-            relay_state=relay_state,
         )
         if binding == BINDING_HTTP_REDIRECT:
             headers = dict(http_args["headers"])
@@ -850,7 +836,7 @@ def localhost_test_idp_acs_post():
             db.session.add(user)
             db.session.commit()
 
-        else :
+        else:
             user.authentication_backend = 'localhost_test_idp'
             user.hash_password(token)
             db.session.add(user)
@@ -858,7 +844,7 @@ def localhost_test_idp_acs_post():
 
         return render_template('idp/idp_login_success.html', token=token), 200
 
-    except (NameError, AttributeError, KeyError) :
+    except (NameError, AttributeError, KeyError):
         return render_template('errors/500.html'), 500
 
 
@@ -872,14 +858,15 @@ def idp_login_auth():
         if email:
             user = check_login(email, token)
             if user:
-                random_token = rndstr()
+                random_token = secrets.token_hex(16)
                 user.hash_password(random_token)
                 db.session.add(user)
                 db.session.commit()
                 return json.dumps({
-                "success": True,
-                'token': random_token,
-                'user': {'username': user.username, 'id': user.id, 'emailid': user.emailid}})
+                    "success": True,
+                    'token': random_token,
+                    'user': {'username': user.username, 'id': user.id, 'emailid': user.emailid}
+                })
             return jsonify({"success": False}), 401
         return jsonify({"success": False}), 401
     except TypeError:
