+++ mslib/msidp/idp_conf.py
@@ -1,8 +1,8 @@
 # -*- coding: utf-8 -*-
 """
 
-    mslib.idp.idp_conf.py
-    ~~~~~~~~~~~~~~~~~~~~~
+    mslib.msidp.idp_conf.py
+    ~~~~~~~~~~~~~~~~~~~~~~~
 
     SAML2 IDP configuration with bindings, endpoints, and authentication contexts.
 
@@ -40,7 +40,8 @@ from saml2.sigver import get_xmlsec_binary
 
 XMLSEC_PATH = get_xmlsec_binary()
 
-# CRTs and metadata files can be generated through the mscolab server. if configured that way CRTs DIRs should be same in both IDP and mscolab server.
+# CRTs and metadata files can be generated through the mscolab server.
+# if configured that way CRTs DIRs should be same in both IDP and mscolab server.
 BASE_DIR = os.path.expanduser("~")
 DATA_DIR = os.path.join(BASE_DIR, "colabdata")
 MSCOLAB_SSO_DIR = os.path.join(DATA_DIR, 'datasso')
@@ -50,12 +51,13 @@ BASEDIR = os.path.abspath(os.path.dirname(__file__))
 
 def full_path(local_file):
     """Return the full path by joining the BASEDIR and local_file."""
-
     return os.path.join(BASEDIR, local_file)
+
 
 def sso_dir_path(local_file):
     """Return the full path by joining the MSCOLAB_SSO_DIR and local_file."""
     return os.path.join(MSCOLAB_SSO_DIR, local_file)
+
 
 HOST = 'localhost'
 PORT = 8088
@@ -73,14 +75,14 @@ SERVER_KEY = f"{MSCOLAB_SSO_DIR}/key_local_idp.key"
 CERT_CHAIN = ""
 SIGN_ALG = None
 DIGEST_ALG = None
-#SIGN_ALG = ds.SIG_RSA_SHA512
-#DIGEST_ALG = ds.DIGEST_SHA512
+# SIGN_ALG = ds.SIG_RSA_SHA512
+# DIGEST_ALG = ds.DIGEST_SHA512
 
 
 CONFIG = {
     "entityid": f"{BASE}/idp.xml",
     "description": "My IDP",
-    #"valid_for": 168,
+    # "valid_for": 168,
     "service": {
         "aa": {
             "endpoints": {
@@ -133,18 +135,11 @@ CONFIG = {
             "policy": {
                 "default": {
                     "lifetime": {"minutes": 15},
-                    "attribute_restrictions": None, # means all I have
+                    "attribute_restrictions": None,  # means all I have
                     "name_form": NAME_FORMAT_URI,
-                    #"entity_categories": ["swamid", "edugain"]
+                    # "entity_categories": ["swamid", "edugain"]
                 },
             },
-            # ToDo refactor, Could we also move the idp/modules to the colabdata? For what are they needed?
-            # see discussion in https://github.com/Open-MSS/MSS/pull/1818#pullrequestreview-1554358384
-
-            # ToDo refactor,  Is this token needed ? see discussion
-            # in https://github.com/Open-MSS/MSS/pull/1818#issuecomment-1658030366
-
-            "subject_data": "./idp.subject",
             "name_id_format": [NAMEID_FORMAT_TRANSIENT,
                                NAMEID_FORMAT_PERSISTENT]
         },
@@ -175,7 +170,7 @@ CONFIG = {
     # This database holds the map between a subject's local identifier and
     # the identifier returned to a SP
     "xmlsec_binary": XMLSEC_PATH,
-    #"attribute_map_dir": "../attributemaps",
+    # "attribute_map_dir": "../attributemaps",
     "logging": {
         "version": 1,
         "formatters": {
@@ -205,15 +200,11 @@ CONFIG = {
     },
 }
 
-# Authentication contexts
-
-    #(r'verify?(.*)$', do_verify),
-
 CAS_SERVER = "https://cas.umu.se"
 CAS_VERIFY = f"{BASE}/verify_cas"
 PWD_VERIFY = f"{BASE}/verify_pwd"
 
 AUTHORIZATION = {
-    "CAS" : {"ACR": "CAS", "WEIGHT": 1, "URL": CAS_VERIFY},
-    "UserPassword" : {"ACR": "PASSWORD", "WEIGHT": 2, "URL": PWD_VERIFY}
+    "CAS": {"ACR": "CAS", "WEIGHT": 1, "URL": CAS_VERIFY},
+    "UserPassword": {"ACR": "PASSWORD", "WEIGHT": 2, "URL": PWD_VERIFY}
 }
