+++ mslib/msidp/templates/root.mako
